// src/app/shared/models/commons/notificationMessage.ts
export class NotificationMessage {
    id?: number
    name?: string;
    title?: string;
    description?: string;
    purpose?: string;
    messageType?: string;
    jsonData?: string;
    status?: string;
    createdBy?: string;
    modifiedBy?: string;
    createdDate?: Date;
    modifiedDate?: Date;
}